requirements.txt
Django>=4.0
google-generativeai
chromadb
pypdf
python-dotenv